requirements.txt
# AI Player - Core Dependencies
# ML/RL Framework
torch>=2.0.0
torchvision>=0.15.0
stable-baselines3>=2.0.0
gymnasium>=0.28.0

# Computer Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0

# macOS Native Tools
pyobjc-framework-Cocoa>=9.0
pyobjc-framework-Quartz>=9.0
pyobjc-framework-CoreGraphics>=9.0
pyobjc-framework-Vision>=9.0

# OCR (macOS Native)
ocrmac>=1.2.0

# Input Automation (fallback if needed)
pyautogui>=0.9.54

# Logging & Monitoring
tensorboard>=2.13.0
wandb>=0.15.0

# Data Management
pandas>=2.0.0
matplotlib>=3.7.0

# Utilities
tqdm>=4.65.0
pydantic>=2.0.0
pyyaml>=6.0.0